+++ hydxlib/tests/test_threedi.py
@@ -5,6 +5,8 @@ from hydxlib.threedi import check_if_element_is_created_with_same_code
 from hydxlib.threedi import get_cross_section_details
 from hydxlib.threedi import get_hydx_default_profile
 from hydxlib.threedi import get_mapping_value
+from hydxlib.threedi import is_closed
+from hydxlib.threedi import make_open
 from hydxlib.threedi import Threedi
 from unittest import mock
 
@@ -69,6 +71,7 @@ def test_import_hydx_unknown_connection_types(caplog):
     hydx.connections = [
         mock.Mock(identificatieknooppuntofverbinding="ovs82", typeverbinding="XXX")
     ]
+    hydx.profiles = []
     threedi = Threedi()
     threedi.import_hydx(hydx)
     assert '"typeverbinding" is not recognized' in caplog.text
@@ -83,6 +86,7 @@ def test_structure_does_not_exist_error(caplog):
     hydx.connections = [
         mock.Mock(identificatieknooppuntofverbinding="pmp1", typeverbinding="PMP")
     ]
+    hydx.profiles = []
     threedi = Threedi()
     threedi.import_hydx(hydx)
     assert "Structure does not exist for connection" in caplog.text
@@ -140,13 +144,7 @@ def test_import_hydx(hydx):
         "material": 0,
         "sewerage_type": 0,
         "calculation_type": 1,
-        "cross_section_details": {
-            "code": "round_1.1",
-            "shape": 2,
-            "width": 1.1,
-            "height": None,
-        },
-        "cross_section_code": "round_1.1",
+        "cross_section_code": "BET1100",
     }
     assert threedi.impervious_surfaces[0] == {
         "code": "1",
@@ -179,31 +177,24 @@ def test_import_hydx(hydx):
         "display_name": "1009-1009-1",
         "start_node.code": "knp8",
         "end_node.code": "knp55",
-        "cross_section_details": {
-            "code": "rectangle_w3.0_open",
-            "shape": 1,
-            "width": 3.0,
-            "height": None,
-        },
         "crest_type": 4,
         "crest_level": 2.7,
         "discharge_coefficient_positive": 0.9,
         "discharge_coefficient_negative": 0.9,
         "sewerage": True,
-        "cross_section_code": "rectangle_w3.0_open",
+        "cross_section_code": "weir_ovs83",
     }
     assert threedi.orifices[1] == {
         "code": "drl97",
         "display_name": "2002-2002-1",
         "start_node.code": "knp16",
         "end_node.code": "knp60",
-        "cross_section_details": {"shape": 2, "width": 0.4, "height": None},
         "discharge_coefficient_positive": 0,
         "discharge_coefficient_negative": 0.8,
         "sewerage": True,
         "crest_type": 4,
         "crest_level": 0.0,
-        "cross_section_code": "round_0.4",
+        "cross_section_code": "PVC400",
     }
 
 
@@ -220,35 +211,92 @@ def get_profile(**kwargs):
         ("RND", 500, None, {"shape": 2, "width": 0.5, "height": None}),
         ("EIV", 1100, None, {"shape": 3, "width": 1.1, "height": None}),
         ("RHK", 800, 500, {"shape": 0, "width": 0.8, "height": 0.5}),
+        ("EIG", 1100, None, {"shape": 8, "width": 1.1, "height": None}),
+        ("TPZ", 800, 400, {"shape": 6, "width": "0.8 1.6 0", "height": "0 0.4 0.4"}),
     ],
 )
 def test_get_cross_section_details(vrm, bre, hgt, expected):
     profiel = get_profile(
-        vormprofiel=vrm, breedte_diameterprofiel=bre, hoogteprofiel=hgt
+        identificatieprofieldefinitie="PRO",
+        vormprofiel=vrm,
+        breedte_diameterprofiel=bre,
+        hoogteprofiel=hgt,
+        materiaal="PVC",
+        tabulatedbreedte="",
+        tabulatedhoogte="",
     )
     actual = get_cross_section_details(profiel, None, None)
+    expected.setdefault("code", "PRO")
+    expected.setdefault("material", 1)
     assert actual == expected
 
 
 @pytest.mark.parametrize(
     "vrm,expected_shape",
     [
-        ("TAB", 5),
-        ("HEU", 6),
-        ("MVR", 6),
-        ("UVR", 6),
-        ("OVA", 6),
+        ("TAB", 7),
+        ("HEU", 7),
+        ("MVR", 7),
+        ("UVR", 7),
+        ("OVA", 7),
+        ("TPZ", 7),
+        ("YZP", 7),
     ],
 )
 def test_get_cross_section_details_tabulated(vrm, expected_shape):
     profiel = get_profile(
+        identificatieprofieldefinitie="PRO",
         vormprofiel=vrm,
         tabulatedbreedte="0.1 0.5 1 1.5",
         tabulatedhoogte="0 0.25 0.5 1",
+        materiaal="PVC",
     )
     actual = get_cross_section_details(profiel, None, None)
     assert actual == {
+        "code": "PRO",
         "shape": expected_shape,
         "width": profiel.tabulatedbreedte,
         "height": profiel.tabulatedhoogte,
+        "material": 1,
     }
+
+
+@pytest.mark.parametrize(
+    "cross_section,expected",
+    [
+        ({"shape": 0, "width": 0.5, "height": 0.3}, True),
+        ({"shape": 1, "width": 0.5, "height": 0.3}, True),
+        ({"shape": 2, "width": 0.5, "height": 0.3}, True),
+        ({"shape": 3, "width": 0.5, "height": 0.3}, True),
+        ({"shape": 8, "width": 0.5, "height": 0.3}, True),
+        ({"shape": 7, "width": "0 1 2", "height": "0.5 0 0.5"}, False),
+        ({"shape": 7, "width": "0 1 2 1 0", "height": "0.5 0 0.5 1.0 0.5"}, True),
+        ({"shape": 6, "width": "0 1 0", "height": "0 1 1"}, True),
+        ({"shape": 6, "width": "0 1", "height": "0 1"}, False),
+    ],
+)
+def test_cross_section_is_closed(cross_section, expected):
+    assert is_closed(cross_section) == expected
+
+
+def test_cross_section_make_open():
+    cross_section = {"shape": 6, "width": "0 1 0", "height": "0 1 1"}
+    make_open(cross_section)
+    assert cross_section == {"shape": 6, "width": "0 1", "height": "0 1"}
+
+
+@pytest.mark.parametrize(
+    "cross_section",
+    [
+        {"shape": 0, "width": 0.5, "height": 0.3},
+        {"shape": 1, "width": 0.5, "height": 0.3},
+        {"shape": 2, "width": 0.5, "height": 0.3},
+        {"shape": 3, "width": 0.5, "height": 0.3},
+        {"shape": 8, "width": 0.5, "height": 0.3},
+        {"shape": 7, "width": "0 1 2", "height": "0.5 0 0.5"},
+        {"shape": 7, "width": "0 1 2 1 0", "height": "0.5 0 0.5 1.0 0.5"},
+    ],
+)
+def test_cross_section_make_open_err(cross_section):
+    with pytest.raises(ValueError):
+        assert make_open(cross_section)